refactor: database/db.go to config/database.go and remove useless test case
